requirements.txt
addict==2.4.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.0
ccimport==0.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.25.0
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
cumm-cu116==0.4.9
cycler==0.11.0
dash==2.14.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
develop==0.2.1
docopt==0.6.2
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
fire==0.6.0
Flask==3.0.0
fonttools==4.51.0
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.4
lark==1.1.5
lazy_loader==0.4
lit==15.0.7
llvmlite==0.40.0
lyft-dataset-sdk==0.0.8
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.7
mdurl==0.1.2
MinkowskiEngine @ git+https://github.com/NVIDIA/MinkowskiEngine.git@02fc608bea4c0549b0a7b00ca1bf15dee4a0b228
mistune==3.0.2
mmcv==2.0.0
mmdet==3.0.0
-e git+https://github.com/open-mmlab/mmdetection3d.git@22aaa47fdb53ce1870ff92cb7e3f96ae38d17f61#egg=mmdet3d
mmengine==0.7.3
mmsegmentation==1.0.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nh3==0.2.17
ninja==1.11.1
notebook==7.1.3
notebook_shim==0.2.4
numba==0.57.0
numpy==1.24.1
nuscenes-devkit==1.1.10
open3d==0.17.0
opencv-python==4.7.0.72
overrides==7.7.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.4
pccm==0.4.7
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.18.0
plyfile==1.0.2
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.10.4
pycocotools==2.0.6
pycparser==2.22
Pygments==2.17.2
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
python-hglib==2.6.2
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.0
qtconsole==5.5.1
QtPy==2.4.1
readme_renderer==43.0
referencing==0.34.0
repo==0.3.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.18.0
ruamel.showoutput==0.1.7
ruff==0.4.4
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.3
Shapely==1.8.5
SharedArray==3.2.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spconv-cu116==2.3.6
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
termcolor==2.3.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1+cu118
torch_scatter==2.1.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trimesh==3.21.6
triton==2.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
yapf==0.33.0
yarg==0.1.9
zipp==3.17.0
